refactor(ui/error): queryErrorBoundary 공통 ui 로 변경
Resolves: #212
